Fix unsigned long/long long mismatch on macOS
VkDeviceSize and size_t mismatch on 64-bit here, unlike Linux/Windows.

+++ src/niagara.cpp
@@ -762,7 +762,7 @@ void buildBLAS(VkDevice device, const std::vector<Mesh>& meshes, const Buffer& v
 
 		totalAccelerationSize = (totalAccelerationSize + sizeInfo.accelerationStructureSize + kAlignment - 1) & ~(kAlignment - 1);
 
-		maxScratchSize = std::max(maxScratchSize, sizeInfo.buildScratchSize);
+		maxScratchSize = std::max(maxScratchSize, size_t(sizeInfo.buildScratchSize));
 	}
 
 	createBuffer(blasBuffer, device, memoryProperties, totalAccelerationSize, VK_BUFFER_USAGE_ACCELERATION_STRUCTURE_STORAGE_BIT_KHR | VK_BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT);
